- breaking out the game state data access layers a bit. Ships are still with fleets, and squads with units; I may break them down more later
