+++ Assets/Scripts/Helpers/Database/FleetDataAccess.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a1f705ddfde65de4fb98fb2bd5dc55d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
