+++ Assets/Scripts/Helpers/Database/PlanetDataAccess.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f5f4e64507fac5408afdd90b906fd32
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
