+++ Assets/Scripts/Helpers/Database/UnitDataAccess.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eee0e8f95f36e3144ba9e38153e97458
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
