+++ Assets/StreamingAssets/GameData/SaveStructure.sql.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 733b9518c35b7fc41a3c256dabf01cab
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
